+++ lib/application/use_cases/media_tasks/pause_all_media_task_executions_use_case.dart
@@ -1,0 +1,11 @@
+import 'package:framelean/application/services/execution/ffmpeg_task_queue_runner.dart';
+
+class PauseAllMediaTaskExecutionsUseCase {
+  final FfmpegTaskQueueRunner queueRunner;
+
+  const PauseAllMediaTaskExecutionsUseCase({required this.queueRunner});
+
+  Future<FfmpegQueueStartResult> call() {
+    return queueRunner.pauseAllRunningTasks();
+  }
+}